backend/requirements.txt
fastapi>=0.110.0
mangum>=0.17.0
boto3>=1.34.0
pydantic>=2.0.0
pydantic[email]
httpx>=0.27.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
